# Implement NotImplemented fallback kernel and use it to replace usage in VariableType