--- a/tools/autograd/gen_inplace_or_view_type.py
+++ b/tools/autograd/gen_inplace_or_view_type.py
@@ -124,6 +124,12 @@ m.impl("${unqual_operator_name_with_overload}",
 );
 """)
 
+AUTOGRAD_NOT_IMPLEMENTED_REGISTRATION = CodeTemplate("""\
+m.impl("${unqual_operator_name_with_overload}",
+   torch::CppFunction::makeFromBoxedFunction<&torch::autograd::autogradNotImplementedFallback>()
+);
+""")
+
 INPLACE_REDISPATCH = CodeTemplate("""\
 {
   at::AutoDispatchBelowADInplaceOrView guard;