--- a/torch/csrc/autograd/custom_function.cpp
+++ b/torch/csrc/autograd/custom_function.cpp
@@ -29,8 +29,9 @@ Variable VariableInfo::zeros(at::OptionalDeviceGuard& device_guard) const {
   }
 }
 
+template <typename F>
 void _foreach_tensor(
-    const std::function<void (size_t, size_t, const at::Tensor&)>& fn,
+    F fn,
     torch::jit::Stack* stack,
     size_t stack_start,
     size_t size) {
@@ -42,8 +43,9 @@ void _foreach_tensor(
       const auto& tensor = ivalue.toTensor();
       fn(idx, i, tensor);
       idx++;
-    } if (ivalue.isTensorList()) {
-      for (const auto& tensor : ivalue.toTensorVector()) {
+    } else if (ivalue.isTensorList()) {
+      for (const auto& iv : ivalue.toListRef()) {
+        const auto& tensor = iv.toTensor();
         fn(idx, i, tensor);
         idx++;
       }
@@ -61,57 +63,39 @@ void autogradNotImplementedFallback(const c10::OperatorHandle& op, DispatchKeySe
   const auto num_returns = returns.size();
   const auto stack_start = stack->size() - num_arguments;
   const bool grad_mode = GradMode::is_enabled();
-  std::vector<at::Tensor> tensors_requiring_grad;
+  std::vector<const at::Tensor*> tensors_requiring_grad;
 
-  bool any_is_inplace = false;
-  for (const auto i : c10::irange(num_arguments)) {
-    const auto& alias_info = arguments[i].alias_info();
-    any_is_inplace |= alias_info.has_value() && alias_info->isWrite();
-  }
   // Keep track of which outputs are output of in-place modification
   // so we can rebase_history if necessary
   std::vector<bool> is_inplace_output;
   std::vector<bool> is_aliased_output;
   is_inplace_output.reserve(num_returns);
+  is_aliased_output.reserve(num_returns);
   for (const auto i : c10::irange(num_returns)) {
     const auto& alias_info = returns[i].alias_info();
     is_inplace_output.push_back(alias_info.has_value() && alias_info->isWrite());
     is_aliased_output.push_back(alias_info.has_value());
   }
   size_t num_tensor_inputs = 0;  // Only used for DEBUG-only checks
-  for (const auto i : c10::irange(num_arguments)) {
-    auto& ivalue = (*stack)[stack_start + i];
-    if (ivalue.isTensor()) {
-      // ivalues for optional tensors have isTensor = true
-      const auto& tensor = VariableType::unpack(ivalue.toTensor(), op_name.c_str(), i);
-      if (grad_mode && tensor.requires_grad()) {
-        tensors_requiring_grad.push_back(tensor);
-      }
-      num_tensor_inputs++;
-      TORCH_CHECK_NOT_IMPLEMENTED(!generated::details::isFwGradDefined(tensor), "Trying to use forward AD with ", op_name, " that does not support it.");
-    } else if (ivalue.isTensorList()) {
-      const auto& tensors = VariableType::unpack(ivalue.toTensorVector(), op_name.c_str(), i);
-      for (const auto& tensor : tensors) {
-        if (grad_mode && tensor.requires_grad()) {
-          tensors_requiring_grad.push_back(tensor);
-        }
-        num_tensor_inputs++;
-        TORCH_CHECK_NOT_IMPLEMENTED(!generated::details::isFwGradDefined(tensor), "Trying to use forward AD with ", op_name, " that does not support it.");
-      }
+
+  _foreach_tensor([&](size_t _, size_t idx_arg, const at::Tensor& t) {
+    TORCH_CHECK(t.defined(), "Expected argument ", idx_arg, " of ", op_name, " to be defined.");
+    if (grad_mode && t.requires_grad()) {
+      tensors_requiring_grad.push_back(&t);
     }
-  }
+    num_tensor_inputs++;
+    TORCH_CHECK_NOT_IMPLEMENTED(!generated::details::isFwGradDefined(t), "Trying to use forward AD with ", op_name, " that does not support it.");
+  }, stack, stack_start, num_arguments);
+
   const bool any_requires_grad = tensors_requiring_grad.size() > 0;
 
-  if (any_is_inplace) {
-    for (const auto i : c10::irange(num_arguments)) {
-      auto& ivalue = (*stack)[stack_start + i];
-      if (ivalue.isTensor()) {
-        check_inplace(ivalue.toTensor(), any_requires_grad);
-      } else if (ivalue.isTensorList()) {
-        check_inplace(ivalue.toTensorVector(), any_requires_grad);
-      }
+  _foreach_tensor([&](size_t _, size_t i, const at::Tensor& t) {
+    const auto& alias_info = arguments[i].alias_info();
+    if (alias_info.has_value() && alias_info->isWrite()) {
+      check_inplace(t, any_requires_grad);
     }
-  }
+  }, stack, stack_start, num_arguments);
+
   std::shared_ptr<NotImplemented> grad_fn;
   if (any_requires_grad) {
     grad_fn = std::shared_ptr<NotImplemented>(new NotImplemented(op_name), deleteNode);
@@ -158,14 +142,16 @@ void autogradNotImplementedFallback(const c10::OperatorHandle& op, DispatchKeySe
   #endif
   const auto& ret = torch::jit::last(stack, num_returns);
 
-  for (const auto return_idx : c10::irange(ret.size())) {
-    if (ret[return_idx].isTensor()) {
-      const auto& t = ret[return_idx].toTensor();
-      if (grad_fn && isDifferentiableType(t.scalar_type())) {
-        if (is_inplace_output[return_idx]) {
-          rebase_history(const_cast<at::Tensor&>(t), grad_fn);
-        } else {
-          set_history(const_cast<at::Tensor&>(t), grad_fn);
+  if (any_requires_grad) {
+    for (const auto return_idx : c10::irange(ret.size())) {
+      if (ret[return_idx].isTensor()) {
+        const auto& t = ret[return_idx].toTensor();
+        if (isDifferentiableType(t.scalar_type())) {
+          if (is_inplace_output[return_idx]) {
+            rebase_history(const_cast<at::Tensor&>(t), grad_fn);
+          } else {
+            set_history(const_cast<at::Tensor&>(t), grad_fn);
+          }
         }
       }
     }