Fixes #62032 

We can also do something similar with templates (see [here](https://gist.github.com/ccfedd1eae450815437915f2487dc265
)).
 - The main appeal for the templated version is that we expect it to be faster (currently it is ~1.5x faster, but the boxed kernel also currently does several more checks)

Benefits of using boxed version of 
 - better UX: users do not need to instantiate their own template with the types of the arguments and returns
 - writing checks is easier because we access to alias info (not really sure how we would do this otherwise unless we tell user to explicitly pass in alias/mutation information?)
 - Reduce compile time ~1.2x faster or 16% reduction.

Next steps:
 - make this an actual fallback when autograd kernel is not registered (BC-breaking)
 - Two ways of addressing possible breakages (as discussed offline with @ezyang):
   - someone hitting performance issues from boxed fallback should just use inference mode
   - someone writing a composite op but registered as CPU or CUDA should actually register as composite
 - With the debug checks, there are more possible breakages unfortunately. The fix would be dependent
   specific case 

### Compile time improvements after replacing codegened kernels
Before:
```
Time               LOC
22.388195991516113 9382
22.91336464881897 9522
22.35034680366516 9543
23.813859462738037 10219
22.55459690093994 9258
total time to compile: 114.02036380767822, total loc: 47924
```
After:
```
Time               LOC
26.092895030975342 11861
27.48097252845764 12449
27.181522607803345 12481
29.233663082122803 13627
25.77803874015808 11544
total time to compile: 135.7670919895172, total loc: 61962

```
### Benchmarking
<details>
<summary>
Latest benchmarking

[repo to replicate](https://github.com/soulitzer/autograd-not-implemented-benchmarking)
</summary>

```----------- Baseline -----------
2021-08-11T19:51:25-04:00
Running ./example-app-baseline
Run on (32 X 4238.08 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x16)
  L1 Instruction 32 KiB (x16)
  L2 Unified 512 KiB (x16)
  L3 Unified 16384 KiB (x4)
Load Average: 1.28, 0.50, 0.57
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
op_call_no_requires_grad        688 ns          688 ns       999897
op_call                         679 ns          679 ns      1019086
op_call_boxed                   738 ns          739 ns       945178
----------- Templated -----------
2021-08-11T19:51:27-04:00
Running ./example-app-templated
Run on (32 X 3606.59 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x16)
  L1 Instruction 32 KiB (x16)
  L2 Unified 512 KiB (x16)
  L3 Unified 16384 KiB (x4)
Load Average: 1.26, 0.51, 0.57
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
op_call_no_requires_grad        839 ns          839 ns       816023
op_call                        1649 ns         1649 ns       421783
op_call_boxed                  1705 ns         1706 ns       408794
----------- Boxed -----------
2021-08-11T19:51:30-04:00
Running ./example-app-boxed
Run on (32 X 3596.95 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x16)
  L1 Instruction 32 KiB (x16)
  L2 Unified 512 KiB (x16)
  L3 Unified 16384 KiB (x4)
Load Average: 3.72, 1.04, 0.74
-------------------------------------------------------------------
Benchmark                         Time             CPU   Iterations
-------------------------------------------------------------------
op_call_no_requires_grad        883 ns          883 ns       772852
op_call                        1663 ns         1663 ns       418360
op_call_boxed                  1662 ns         1663 ns       421296

```
</details>

#### Old results for comparison
[code](https://gist.github.com/e74939f2c3993a11555909135fd2ef51)
Boxed kernel:
```
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
op_requires_grad_no_dispatch          1568 ns         1567 ns       441375
op_no_requires_grad_no_dispatch        838 ns          838 ns       844149
op_no_requires_grad_call               987 ns          987 ns       699862
op_requires_grad_call                 1776 ns         1776 ns       391280
op_requires_grad_call_boxed           1776 ns         1776 ns       387758
```
Templated kernel:
```
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
op_requires_grad_no_dispatch          1523 ns         1523 ns       451260
op_no_requires_grad_no_dispatch        787 ns          787 ns       885251
op_no_requires_grad_call               772 ns          772 ns       918897
op_requires_grad_call                 1108 ns         1108 ns       630075
op_requires_grad_call_boxed           1179 ns         1179 ns       594634
```