@ezyang 2021-08-09
This looks like a really good start. I found a bunch of performance inefficiencies in your boxed version which should reduce the 1.5x gap. At a 1.5x gap, I'm not sure I would like to have boxed fallback used for everything in derivatives.yaml, but if we can reduce the gap some more I think what you do in this PR (stop codegen) is very plausible.