@ezyang 2021-08-12
Can you also benchmark what the code generated NotImplemented implementation used to be? Just to make sure we aren't accidentally regressing...